+++ qa/TL0_jupyter/test_cupy.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 # used pip packages
-pip_packages="jupyter numpy matplotlib cupy"
+pip_packages="jupyter numpy matplotlib cupy imageio"
 target_dir=./docs/examples
 
 test_body() {
@@ -8,7 +8,7 @@ test_body() {
         "custom_operations/gpu_python_operator.ipynb"
         "general/data_loading/external_input.ipynb"
     )
-    for f in ${test_files[@]};
+    for f in ${test_files[@]}; do
         jupyter nbconvert --to notebook --inplace --execute \
                         --ExecutePreprocessor.kernel_name=python${PYVER:0:1} \
                         --ExecutePreprocessor.timeout=300 $f;
