+++ qa/TL0_jupyter/test_nofw.sh
@@ -15,7 +15,7 @@ test_body() {
     # test all jupyters except one related to a particular FW,
     # and one requiring a dedicated HW (multiGPU and OF)
     # optical flow requires TU102 architecture whilst this test can be run on any GPU
-    black_list_files="multigpu\|mxnet*\|tensorflow*\|pytorch*\|paddle*\|*external_input.ipynb\|optical_flow*\|python_operator*\|#"
+    black_list_files="multigpu\|mxnet\|tensorflow\|pytorch\|paddle\|external_input.ipynb\|optical_flow\|python_operator\|#"
 
     find * -name "*.ipynb" | sed "/${black_list_files}/d" | xargs -i jupyter nbconvert \
                     --to notebook --inplace --execute \
